+++ CHANGELOG.md
@@ -12,6 +12,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- hnefatafl-client: you have to have a username to press "Login" or "Create
+  Account:.
+
 ### Changed
 
 - cargo update.