+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [6.2.3] - 2026-08-08
+
 ### Fixed
 
 - hnefatafl-server-full - set a delay of 1s, 2s, 4s, 8s, 16s, 30m, for an IP
@@ -941,7 +943,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - hnefatafl-client: Make users and games scrollable.
 - Make all the features dependencies.
 
-[unreleased]: https://codeberg.org/dcampbell/hnefatafl/compare/v6.2.2-1...main
+[unreleased]: https://codeberg.org/dcampbell/hnefatafl/compare/v6.2.3-1...main
+[6.2.3]: https://codeberg.org/dcampbell/hnefatafl/compare/v6.2.2-1...v6.2.3-1
 [6.2.2]: https://codeberg.org/dcampbell/hnefatafl/compare/v6.2.1-1...v6.2.2-1
 [6.2.1]: https://codeberg.org/dcampbell/hnefatafl/compare/v6.2.0-1...v6.2.1-1
 [6.2.0]: https://codeberg.org/dcampbell/hnefatafl/compare/v6.1.1-1...v6.2.0-1